feat(unwind): Remove memory module
