+++ symbolic-unwind/src/lib.rs
@@ -1,6 +1,5 @@
 //! Stack unwinding functionality for `symbolic`.
 #![warn(missing_docs)]
 
-mod base;
+pub mod base;
 pub mod evaluator;
-mod memory;
